final version with floatNumbers

+++ laboratories/FirstLaboratory/app/src/main/java/com/example/firstlaboratory/MainActivity.java
@@ -27,8 +27,8 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 try {
-                    int first = Integer.parseInt(op1.getText().toString());
-                    int second = Integer.parseInt(op2.getText().toString());
+                    float first = Float.parseFloat(op1.getText().toString());
+                    float second = Float.parseFloat(op2.getText().toString());
                     Intent finalResult = new Intent(MainActivity.this, ResultActivity.class);
                     finalResult.putExtra("Result", first + second);
                     startActivity(finalResult);
@@ -44,8 +44,8 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 try {
-                    int first = Integer.parseInt(op1.getText().toString());
-                    int second = Integer.parseInt(op2.getText().toString());
+                    float first = Float.parseFloat(op1.getText().toString());
+                    float second = Float.parseFloat(op2.getText().toString());
                     Intent finalResult = new Intent(MainActivity.this, ResultActivity.class);
                     finalResult.putExtra("Result", first - second);
                     startActivity(finalResult);
@@ -61,8 +61,8 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 try {
-                    int first = Integer.parseInt(op1.getText().toString());
-                    int second = Integer.parseInt(op2.getText().toString());
+                    float first = Float.parseFloat(op1.getText().toString());
+                    float second = Float.parseFloat(op2.getText().toString());
                     Intent finalResult = new Intent(MainActivity.this, ResultActivity.class);
                     finalResult.putExtra("Result", first * second);
                     startActivity(finalResult);
@@ -78,8 +78,8 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 try {
-                    int first = Integer.parseInt(op1.getText().toString());
-                    int second = Integer.parseInt(op2.getText().toString());
+                    float first = Float.parseFloat(op1.getText().toString());
+                    float second = Float.parseFloat(op2.getText().toString());
                     Intent finalResult = new Intent(MainActivity.this, ResultActivity.class);
                     finalResult.putExtra("Result", first / second);
                     startActivity(finalResult);
